refactor(batch): promote map_project_row to shared mappers, tighten batch visibility
Move map_project_row from batch.rs into database/mappers.rs so both
get_project_by_uuid and get_projects_batch share a single authoritative
mapping; removes silent-divergence risk between single-fetch and batch
paths. Change `pub mod batch` to `pub(crate) mod batch` — the module
exposes no public items of its own.
