+++ libs/things3-core/src/batch.rs
@@ -21,14 +21,13 @@
 
 use std::collections::HashSet;
 
-use chrono::{DateTime, Utc};
 use sqlx::{sqlite::SqliteRow, Row, SqlitePool};
 use uuid::Uuid;
 
-use crate::database::mappers::{map_task_row, parse_optional_uuid, parse_uuid_with_fallback};
-use crate::database::{safe_timestamp_convert, ThingsDatabase};
+use crate::database::mappers::{map_project_row, map_task_row};
+use crate::database::ThingsDatabase;
 use crate::error::{Result as ThingsResult, ThingsError};
-use crate::models::{Project, Task, TaskStatus};
+use crate::models::{Project, Task};
 
 /// Conservative chunk size — keeps each round-trip well below SQLite's
 /// `SQLITE_LIMIT_VARIABLE_NUMBER` floor (999) so callers can pass arbitrarily
@@ -148,43 +147,6 @@ where
         }
     }
     Ok(out)
-}
-
-/// Map a `TMTask` row (where `type = 1`) into a [`Project`].
-///
-/// Lifted from `get_project_by_uuid` (`database/core.rs`) so the batch
-/// variant doesn't depend on internal restructuring of the single-fetch.
-fn map_project_row(row: &SqliteRow) -> Project {
-    Project {
-        uuid: parse_uuid_with_fallback(&row.get::<String, _>("uuid")),
-        title: row.get("title"),
-        status: match row.get::<i32, _>("status") {
-            1 => TaskStatus::Completed,
-            2 => TaskStatus::Canceled,
-            3 => TaskStatus::Trashed,
-            _ => TaskStatus::Incomplete,
-        },
-        area_uuid: parse_optional_uuid(row.get::<Option<String>, _>("area")),
-        notes: row.get("notes"),
-        deadline: row
-            .get::<Option<i64>, _>("deadline")
-            .and_then(|ts| DateTime::from_timestamp(ts, 0))
-            .map(|dt| dt.date_naive()),
-        start_date: row
-            .get::<Option<i64>, _>("startDate")
-            .and_then(|ts| DateTime::from_timestamp(ts, 0))
-            .map(|dt| dt.date_naive()),
-        tags: Vec::new(),
-        tasks: Vec::new(),
-        created: {
-            let ts = safe_timestamp_convert(row.get::<f64, _>("creationDate"));
-            DateTime::from_timestamp(ts, 0).unwrap_or_else(Utc::now)
-        },
-        modified: {
-            let ts = safe_timestamp_convert(row.get::<f64, _>("userModificationDate"));
-            DateTime::from_timestamp(ts, 0).unwrap_or_else(Utc::now)
-        },
-    }
 }
 
 #[cfg(test)]
