+++ libs/things3-core/src/lib.rs
@@ -62,7 +62,7 @@ pub mod error;
 pub mod filter_expr;
 
 #[cfg(feature = "batch-operations")]
-pub mod batch;
+pub(crate) mod batch;
 
 #[cfg(feature = "batch-operations")]
 pub mod cursor;
